restructured portable vector code
