+++ libcrux-ml-kem/src/polynomial.rs
@@ -1,8 +1,8 @@
 use crate::vector::{
-    to_standard_domain, FieldElementTimesMontgomeryR, Operations, FIELD_ELEMENTS_IN_VECTOR,
+    to_standard_domain, Operations, FIELD_ELEMENTS_IN_VECTOR,
 };
 
-pub(crate) const ZETAS_TIMES_MONTGOMERY_R: [FieldElementTimesMontgomeryR; 128] = [
+pub(crate) const ZETAS_TIMES_MONTGOMERY_R: [i16; 128] = [
     -1044, -758, -359, -1517, 1493, 1422, 287, 202, -171, 622, 1577, 182, 962, -1202, -1474, 1468,
     573, -1325, 264, 383, -829, 1458, -1602, -130, -681, 1017, 732, 608, -1542, 411, -205, -1571,
     1223, 652, -552, 1015, -1293, 1491, -282, -1544, 516, -8, -320, -666, -1618, -1162, 126, 1469,
