+++ libcrux-ml-kem/src/vector/portable.rs
@@ -1,25 +1,24 @@
-//! Vectors for libcrux using aarch64 (neon) intrinsics
+use super::Operations;
 
-use super::{Operations, FIELD_MODULUS};
-
-// mod sampling;
 mod vector_type;
 mod arithmetic;
 mod compress;
 mod ntt;
 mod serialize;
+mod sampling;
 
-pub(crate) use vector_type::SIMD128Vector;
 use vector_type::*;
 use arithmetic::*;
 use compress::*;
 use ntt::*;
 use serialize::*;
+use sampling::*;
 
-impl Operations for SIMD128Vector {
-    #[inline(always)]
+pub(crate) use vector_type::PortableVector;
+
+impl Operations for PortableVector {
     fn ZERO() -> Self {
-        ZERO()
+        zero()
     }
 
     fn from_i16_array(array: &[i16]) -> Self {
@@ -43,7 +42,7 @@ impl Operations for SIMD128Vector {
     }
 
     fn shift_right<const SHIFT_BY: i32>(v: Self) -> Self {
-        shift_right::<SHIFT_BY>(v)
+        shift_right::<{ SHIFT_BY }>(v)
     }
 
     fn cond_subtract_3329(v: Self) -> Self {
@@ -54,8 +53,8 @@ impl Operations for SIMD128Vector {
         barrett_reduce(v)
     }
 
-    fn montgomery_multiply_by_constant(v: Self, c: i16) -> Self {
-        montgomery_multiply_by_constant(v, c)
+    fn montgomery_multiply_by_constant(v: Self, r: i16) -> Self {
+        montgomery_multiply_by_constant(v, r)
     }
 
     fn compress_1(v: Self) -> Self {
@@ -70,24 +69,24 @@ impl Operations for SIMD128Vector {
         decompress_ciphertext_coefficient::<COEFFICIENT_BITS>(v)
     }
 
-    fn ntt_layer_1_step(a: Self, zeta1: i16, zeta2: i16, zeta3: i16, zeta4: i16) -> Self {
-        ntt_layer_1_step(a, zeta1, zeta2, zeta3, zeta4)
+    fn ntt_layer_1_step(a: Self, zeta0: i16, zeta1: i16, zeta2: i16, zeta3: i16) -> Self {
+        ntt_layer_1_step(a, zeta0, zeta1, zeta2, zeta3)
     }
 
-    fn ntt_layer_2_step(a: Self, zeta1: i16, zeta2: i16) -> Self {
-        ntt_layer_2_step(a, zeta1, zeta2)
+    fn ntt_layer_2_step(a: Self, zeta0: i16, zeta1: i16) -> Self {
+        ntt_layer_2_step(a, zeta0, zeta1)
     }
 
     fn ntt_layer_3_step(a: Self, zeta: i16) -> Self {
         ntt_layer_3_step(a, zeta)
     }
 
-    fn inv_ntt_layer_1_step(a: Self, zeta1: i16, zeta2: i16, zeta3: i16, zeta4: i16) -> Self {
-        inv_ntt_layer_1_step(a, zeta1, zeta2, zeta3, zeta4)
+    fn inv_ntt_layer_1_step(a: Self, zeta0: i16, zeta1: i16, zeta2: i16, zeta3: i16) -> Self {
+        inv_ntt_layer_1_step(a, zeta0, zeta1, zeta2, zeta3)
     }
 
-    fn inv_ntt_layer_2_step(a: Self, zeta1: i16, zeta2: i16) -> Self {
-        inv_ntt_layer_2_step(a, zeta1, zeta2)
+    fn inv_ntt_layer_2_step(a: Self, zeta0: i16, zeta1: i16) -> Self {
+        inv_ntt_layer_2_step(a, zeta0, zeta1)
     }
 
     fn inv_ntt_layer_3_step(a: Self, zeta: i16) -> Self {
@@ -97,12 +96,12 @@ impl Operations for SIMD128Vector {
     fn ntt_multiply(
         lhs: &Self,
         rhs: &Self,
+        zeta0: i16,
         zeta1: i16,
         zeta2: i16,
         zeta3: i16,
-        zeta4: i16,
     ) -> Self {
-        ntt_multiply(lhs, rhs, zeta1, zeta2, zeta3, zeta4)
+        ntt_multiply(lhs, rhs, zeta0, zeta1, zeta2, zeta3)
     }
 
     fn serialize_1(a: Self) -> [u8; 2] {
@@ -154,32 +153,6 @@ impl Operations for SIMD128Vector {
     }
 
     fn rej_sample(a: &[u8], out: &mut [i16]) -> usize {
-        // FIXME: The code in rejsample fails on the CI machines.
-        // We need to understand why and fix it before using it.
-        // We use the portable version in the meantime.
         rej_sample(a, out)
     }
-}
-
-#[inline(always)]
-pub(crate) fn rej_sample(a: &[u8], result: &mut [i16]) -> usize {
-    let mut sampled = 0;
-    for bytes in a.chunks(3) {
-        let b1 = bytes[0] as i16;
-        let b2 = bytes[1] as i16;
-        let b3 = bytes[2] as i16;
-
-        let d1 = ((b2 & 0xF) << 8) | b1;
-        let d2 = (b3 << 4) | (b2 >> 4);
-
-        if d1 < FIELD_MODULUS && sampled < 16 {
-            result[sampled] = d1;
-            sampled += 1
-        }
-        if d2 < FIELD_MODULUS && sampled < 16 {
-            result[sampled] = d2;
-            sampled += 1
-        }
-    }
-    sampled
 }
