+++ libcrux-ml-kem/src/vector/portable/vector_type.rs
@@ -1,0 +1,25 @@
+use crate::vector::traits::FIELD_ELEMENTS_IN_VECTOR;
+
+/// Values having this type hold a representative 'x' of the Kyber field.
+/// We use 'fe' as a shorthand for this type.
+pub(crate) type FieldElement = i16;
+
+#[derive(Clone, Copy)]
+pub struct PortableVector {
+    pub(crate) elements: [FieldElement; FIELD_ELEMENTS_IN_VECTOR],
+}
+
+#[allow(non_snake_case)]
+#[inline(always)]
+pub fn zero() -> PortableVector {
+    PortableVector {
+        elements: [0i16; FIELD_ELEMENTS_IN_VECTOR],
+    }
+}
+
+#[inline(always)]
+pub fn from_i16_array(array: &[i16]) -> PortableVector {
+    PortableVector {
+        elements: array[0..16].try_into().unwrap(),
+    }
+}
